requirements.txt
Flask>=2.2.2
flask_cors
Annoy
python-dotenv
numpy
transformers==4.31.0
torch
TTS
sentence-transformers
xformers